+++ hydromt/__init__.py
@@ -14,7 +14,7 @@ import warnings  # noqa: F401
 import netCDF4  # noqa: F401
 
 # submodules
-from hydromt import data_catalog, gis, io, model, stats
+from hydromt import data_catalog, gis, model, stats
 
 # high-level methods
 from hydromt._utils.log import initialize_logging
@@ -32,7 +32,6 @@ __all__ = [
     # submodules
     "data_catalog",
     "gis",
-    "io",
     "model",
     "stats",
     # raster and vector accessor